gnlcomposition: Make the pipeline update protected from any child changes
When calling the gnl_object_commit method, we need to make sure that no
object can be changed and, it should be true until the pipeline has
fully been udpated.

We now need the update_pipeline function to be called with the
COMP_OBJECTS_LOCK taken, and we do not unlock that lock during the whole
pipeline updating process.

https://bugzilla.gnome.org/show_bug.cgi?id=701287