gnlcomposition: Remove support for gaps
The way we were handling gap was weird and not natural at all for
users, remove it for now.

In the long term, we should have a proper way of filling gaps in the
API and for now only emit an STREAM_ERROR message on the bus when a
gap is detected.

The user is now responsible for filling gaps in the composition

https://bugzilla.gnome.org/show_bug.cgi?id=701287