gnlcomposition: Drop any buffer flowing from a source when updating pipeline
There is a race where we can have an EOS from the element which
srcpad is our ghost pad but other elements upstream are still outputing
data and we flush the element so its pads have no segment info
anymore, and we end up having warning about flow before segement event.

The solution here is that we do not let any data flowing when we are
updating the pipeline.