nlecomposition: Activate the composition ghostpad to flush downstream
Since commit 060b16ac75ac227d4cfe1db89ccdc4f4b31545ff
"pad: don't accept flush-stop on inactive pads" in -core, the flush_stop event will not be
fowarded downstream in case the pad is not activated. In our case the element is in
READY state, so pads are deactivated. In that commit we simply make sure that the
event can be fowarded downstream